renames state name file to state_name.py
