+++ pgmpy/utils/__init__.py
@@ -1,5 +1,5 @@
 from .mathext import cartesian, sample_discrete
-from .VariableStateNameDecorators import StateNameInit, StateNameDecorator
+from .state_name import StateNameInit, StateNameDecorator
 
 __all__ = ['cartesian',
            'sample_discrete',
